extract change_name_button
